refactor(develop): add special_plugins.go
